+++ src/hooks/useTransactionColumn.tsx
@@ -67,7 +67,6 @@ export const useTransactionColumn = (accountNumber?: string): ColumnsType<any> =
       },
       title: 'Time',
     },
-
     {
       align: 'right',
       dataIndex: 'coins',
